Move collection tests to their own class, remove the dependency on links argument
